For PSP, change typedef unsigned long DWORD; into typedef u32 DWORD;

+++ JGE/JGE/include/JTypes.h
@@ -190,7 +190,7 @@
 	#define FRAME_BUFFER_SIZE		FRAME_BUFFER_WIDTH*SCREEN_HEIGHT*PIXEL_SIZE
 
 	#define SLICE_SIZE_F			64.0f
-	typedef unsigned long DWORD;
+	typedef u32 DWORD;
 
 	#define BLEND_ZERO					0x1000
 	#define BLEND_ONE					0x1002
